factor out the user plugins to their respected folders
