--- lua/core/lambda/conditions.lua
@@ -1,7 +1,0 @@
-function test()
-    print("test")
-end
-
-function pointer()
-    print("this is a pointer function")
-end
