+++ lua/modules/ai/plugins.lua
@@ -6,9 +6,6 @@ ai({
     cmd = "Copilot",
     lazy = true,
     init = function()
-        --[[ vim.opt.completeopt = "menuone,noselect" ]]
-        vim.g.copilot_enabled = lambda.config.sell_your_soul
-        -- Have copilot play nice with nvim-cmp.
         vim.g.copilot_no_tab_map = true
         vim.g.copilot_assume_mapped = true
         vim.g.copilot_tab_fallback = ""
